Reduce some test code duplication
